refactor: Polish identities app to match DRF standards
- Add docstrings to __init__.py, apps.py, views/__init__
- Convert admin.py to admin/ directory package (two admin classes)
- Use custom IsAdminUser permission in both admin viewsets

Made-with: Cursor
