+++ server/apps/identities/__init__.py
@@ -1,0 +1,7 @@
+"""
+Identities app.
+
+Manages user identity definitions — the core coaching constructs that
+represent who a user wants to become. Provides CRUD endpoints, identity
+image generation via AI chat sessions, and PDF export of identity cards.
+"""
