+++ server/apps/identities/views/__init__.py
@@ -1,3 +1,13 @@
+"""
+Identity views.
+
+Exports:
+    IdentityViewSet: Public CRUD and PDF download for authenticated users.
+    IdentityImageChatViewSet: Public image generation chat endpoints.
+    AdminIdentityViewSet: Admin identity operations including image generation.
+    AdminIdentityImageChatViewSet: Admin image chat endpoints for any user.
+"""
+
 from .admin_identity_image_chat_view_set import AdminIdentityImageChatViewSet
 from .admin_identity_view_set import AdminIdentityViewSet
 from .identity_image_chat_view_set import IdentityImageChatViewSet
